feat: Codebase Improvements - Additional Code Testing
test: test calc_pi

test: test calc_dxy

test: FST unit tests (#84)

This PR adds unit tests over FST calculation at biallelic sites in
diploid samples (the "base case"). The tests cover both Hudson and
Weir-Cockerham estimation.

test: Multi-locus tests for `dxy` and `FST` (#87)

This PR adds unit tests for multi-locus genotype arrays for both `dxy`
and `FST`. (There is already such a test for `pi`.)

fix: `bypass_invariant_check` flag (#89)

Closes #81. This PR removes default `bypass_invariant_check` arg on
`PixyArgs` and adds a quick unit-test to make sure our "yes/no"
arguments are transformed correctly by `check_and_validate_args`. We may
not need the test once we address #82 but I wanted to include it in the
meantime.

test: Add regression markers (#92)

This PR marks the regression tests with the `pytest.mark.regression`
marker, so they may be skipped with `pytest -m "not regression"`.

This is intended to reduce the time spent waiting on the test suite
during local development.

Co-authored-by: Matt Stone <matt@fulcrumgenomics.com>
Co-authored-by: Erin McAuley <erin@fulcrumgenomics.com>