feat: Support samples of arbitrary ploidy
tests: add `test_calc_pi_arbitrary_ploidy()` (#83)

This PR adds a test for `calc_pi` in the context of arbitrary ploidy
(tetraploid for now, but we can expand in the future).

tests: add `test_calc_fst_hudson_arbitrary_ploidy()` (#91)

This PR adds `test_calc_fst_hudson_arbitrary_ploidy()` -- based on the
`scikit-allel` implementation, it looks like FST_Hudson works with
arbitrary ploidy. The same cannot be said for FST_WC, whose calculation
in the context of arbitrary ploidy is not supported by `scikit-allel` at
this time.

feat: add `NotImplementedError` if FST_WC is requested in the context of arbitrary ploidy (#93)

This PR adds an extra guardrail in `pixy` if users request the
Weir-Cockerham FST calculation for non-diploid genomes -- `scikit-allel`
does not natively support this feature yet, so it is important to handle
the error gracefully within `pixy`.

Co-authored-by: Matt Stone <matt@fulcrumgenomics.com>
Co-authored-by: Erin McAuley <erin@fulcrumgenomics.com>